Fix UI bug with 'ranges' checkboxes
When unticking all boxes and then ticking one that isn't "day",
it would submit as that one + day.

This was because each time the checkbox-change event handler was
called, it would redo the logic of creating the initial value
from the query parameter and inserting the default as ["day"],
which is only correct the first time.